app/requirements.txt
django==4.2.1
uwsgi==2.0.21
python-dotenv==1.0.0
django-split-settings==1.2.0
django-admin-rangefilter==0.10.0
django-debug-toolbar==4.1.0
psycopg2-binary==2.9.6
django-extensions==3.2.3
django-cors-headers==4.0.0